requirements.txt
discord.py
pyyamlconfig
imgurpython
requests
